pre-commit: add pytype, see rest of commit message for dev setup info
The required setup for this is somewhat awkward. At time of writing
Python 3.9 is not supported by latest pytype, 3.6-3.8 is. But in
addition to that, we need the interpreter corresponding to the
python_version pytype setting, and we need dependencies for this Python
installed for pytype to be able to resolve imports. Ideally that version
would be 3.4, but the latest pytype at the moment only supports >= 3.5,
so make do with that.

So effectively for development, I suggest two pyenv virtualenvs: one
with Python 3.8.x and the requirements-dev deps installed in it, another
with Python 3.5.x and the requirements-test deps installed in it, and
both listed in .pyenv-version, in that order.