ci(github): stop running on schedule
Scheduled runs were originally added as a companion to version specifier
ranges, so that they'd alert if something was broken with possible newer
versions.

Now that we no longer have ranges like that for the linter deps,
breakage of that kind no longer happens with them. Test dependencies do
have inexact versioning, but we rely on rather frequent linter PR's to
work as a scheduler and alert us as a side effect if something with the
test deps breaks.