ci(release): set package-name for release-please
Until the general tooling is ready/ier to move project metadata to
pyproject.toml and not duplicate its info elsewhere, e.g. setup.cfg, set
it here.

By duplicating the package name like this means we are duplicating
anyway, but less than what would be required if we'd move to
pyproject.toml.

Refs https://github.com/googleapis/release-please/issues/1140